Creating Marketplace class to manage MPAC data.
